requirements.txt
ansible>=6.0.0
ansible-core>=2.13.0
jinja2>=3.0.0
PyYAML>=6.0.0
requests>=2.28.0
docker>=6.0.0
kubernetes>=24.0.0
